HU-05 - Implemetación de IA mejorada, falta enviar signos y sintomas a  la base de Datos

+++ Dtos/TriageRequestDto.cs
@@ -9,6 +9,6 @@ public class VitalSignsDto
     public int HeartRate { get; set; }
     public int RespiratoryRate { get; set; }
     public string BloodPressure { get; set; } = string.Empty;
-    public decimal Temperature { get; set; }
+    public double Temperature { get; set; }
     public int OxygenSaturation { get; set; }
 }
